// archive/legacy-review/idefix-backoffice/Idefix Backoffice/gstech-backoffice/src/js/core/api/types.ts
export type FetchApi = (
  url: string,
  config?: RequestInit & {
    params?: { [key: string]: any } | null;
  },
) => Promise<any>;
// ...
export interface SettingsAPI {
  load: () => Promise<Settings>;
  getBrandCountries: (brandId: string) => Promise<CountrySettings[]>;
  getCountries: () => Promise<CountrySettings[]>;
  updateCountry: (
    brandId: string,
    countryId: string,
    countryDraft: CountrySettingsDraft,
  ) => Promise<Omit<CountrySettings, "name">>;
  getGames: () => Promise<GameSettings[]>;
  createGame: (gameDraft: GameDraft, profileDrafts: ProfileDraft[]) => Promise<GameSettings>;
  updateGame: (gameId: number, gameDraft: GameDraft) => Promise<GameSettings>;
  getGameProfiles: (gameId: number) => Promise<GameProfileSetting[]>;
  getAvailableGameProfiles: () => Promise<{
    brandId: string;
    brandName: string;
    availableProfiles: GameProfile[];
  }>;
  updateProfiles: (
    gameId: number,
    profileDrafts: {
      brandId: string;
      gameProfileId: number;
    }[],
  ) => Promise<{
    gameId: number;
    brandId: string;
    gameProfileId: number;
  }>;
  getGameManufacturers: () => Promise<GameManufacturer[]>;
  getGameManufacturer: (gameManufacturerId: string) => Promise<GameManufacturer & { blockedCountries: any[] }>;
  updateGameManufacturer: (
    gameManufacturerId: string,
    manufacturerDraft: Omit<GameManufacturer, "id" | "license" | "name" | "parentId">,
  ) => Promise<OKResponse>;

  getBonuses: (brandId: string) => Promise<Bonus[]>;
  updateBonus: (bonusId: number, bonusDraft: UpdateBonusRequest) => Promise<OKResponse>;
  createBonus: (brandId: string, bonusDraft: CreateBonusRequest) => Promise<Bonus>;
  getBonusLimits: (bonusId: number) => Promise<BonusLimit[]>;
  getAvailableBonusLimits: (brandId: string) => Promise<AvailableBonusLimits>;
  updateBonusLimits: (bonusId: number, bonusLimitDrafts: UpdateBonusLimitsRequest) => Promise<BonusLimit>;

  getPaymentMethods: () => Promise<PaymentMethod[]>;
  getPaymentProviders: (paymentMethodId: number) => Promise<PaymentMethodProvider>;
  updatePaymentProviders: (
    paymentMethodId: number,
    values: Omit<PaymentMethodProvider, "id" | "name" | "paymentProviders">,
  ) => Promise<OKResponse>;
  getPaymentProviderDetails: (paymentProviderId: number) => Promise<PaymentProvider>;
  updatePaymentProviderDetails: (paymentProviderId: number, values: any) => Promise<any>;

  getPromotions: (brandId: string) => Promise<Promotion[]>;
  createPromotion: (promotionDraft: Omit<Promotion, "id"> & { brandId: string }) => Promise<Promotion>;
  updatePromotion: (promotionId: number, promotionDraft: Omit<Promotion, "id">) => Promise<Promotion>;

  getGameProfileSettings: (brandId: string) => Promise<GameProfile[]>;
  createGameProfile: (gameProfileDraft: Omit<GameProfile, "id">) => Promise<GameProfile>;
  updateGameProfile: (gameProfileId: number, gameProfileDraft: Omit<GameProfile, "id">) => Promise<GameProfile>;

  getCurrencies: (brandId: string) => Promise<Currency[]>;
  archiveBonus: (bonusId: number) => Promise<true>;
  archivePromotion: (promotionId: number) => Promise<number[]>;

  getPromotionGames: (promotionId: number) => Promise<number[]>;
  addPromotionGames: (promotionId: number, games: number[]) => Promise<{ promotionId: number; gameId: number }[]>;
  updatePromotionGames: (promotionId: number, games: number[]) => Promise<number[]>;

  getRisks: (params?: { manualTrigger: boolean }) => Promise<Risk[]>;
  addRisk: (riskDraft: RiskDraft) => Promise<Risk>;
  updateRisk: (riskId: number, riskDraft: RiskDraft) => Promise<Risk>;
}